+++ noodles-vcf/src/header/record/value/collection.rs
@@ -1,3 +1,7 @@
+//! VCF header record value collection.
+
+use std::{error, fmt};
+
 use indexmap::IndexMap;
 
 use super::Value;
@@ -26,13 +30,48 @@ impl Collection {
         }
     }
 
-    pub(crate) fn add(&mut self, value: Value) {
+    pub(crate) fn add(&mut self, value: Value) -> Result<(), AddError> {
         match (self, value) {
-            (Self::Unstructured(list), Value::String(s)) => list.push(s),
+            (Self::Unstructured(list), Value::String(s)) => {
+                list.push(s);
+                Ok(())
+            }
+            (Self::Unstructured(_), Value::Map(..)) => Err(AddError::TypeMismatch {
+                actual: "structured",
+                expected: "unstructured",
+            }),
             (Self::Structured(map), Value::Map(id, m)) => {
                 map.insert(id, m);
+                Ok(())
             }
-            (_, _) => panic!("invalid"),
+            (Self::Structured(_), Value::String(_)) => Err(AddError::TypeMismatch {
+                actual: "unstructured",
+                expected: "structured",
+            }),
+        }
+    }
+}
+
+/// An error returned when a value fails be added to a collection.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum AddError {
+    /// The value does not match the collection type.
+    TypeMismatch {
+        /// The value type.
+        actual: &'static str,
+        /// The collection type.
+        expected: &'static str,
+    },
+}
+
+impl error::Error for AddError {}
+
+impl fmt::Display for AddError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::TypeMismatch { actual, expected } => {
+                write!(f, "type mismatch: expected {expected}, got {actual}")
+            }
         }
     }
 }